+++ services/web/server/src/simcore_service_webserver/licenses/settings.py
@@ -1,25 +1,9 @@
 from typing import Annotated
 
-from pydantic import AfterValidator, Field, HttpUrl
+from pydantic import Field
 from settings_library.base import BaseCustomSettings
 
-
-def _validate_url_contains_category(url: str) -> str:
-    if "{category}" not in url:
-        msg = "URL must contain '{category}'"
-        raise ValueError(msg)
-    return url
-
-
-class ItisVipSettings(BaseCustomSettings):
-    ITIS_VIP_API_URL: Annotated[str, AfterValidator(_validate_url_contains_category)]
-    ITIS_VIP_CATEGORIES: list[str]
-
-    def get_urls(self) -> list[HttpUrl]:
-        return [
-            HttpUrl(self.ITIS_VIP_API_URL.format(category=category))
-            for category in self.ITIS_VIP_CATEGORIES
-        ]
+from ._itis_vip_settings import ItisVipSettings
 
 
 class LicensesSettings(BaseCustomSettings):
